requirements.txt
streamlit
pandas
google-cloud-bigquery
google-cloud-aiplatform

